+++ src/configuration/management.rs
@@ -8,11 +8,13 @@ const HEALTH_ENDPOINT: &str = "/health";
 const LIVE_ENDPOINT: &str = "/live";
 const READY_ENDPOINT: &str = "/ready";
 const METRICS_ENDPOINT: &str = "/metrics";
+const VERSION_ENDPOINT: &str = "/version";
 
 const HEALTH_PTR: &str = "/health";
 const LIVE_PTR: &str = "/live";
 const READY_PTR: &str = "/ready";
 const METRICS_PTR: &str = "/metrics";
+const VERSION_PTR: &str = "/version";
 
 #[derive(Debug, Default, Clone, Deserialize)]
 /// [`Management`](https://github.com/elefant-dev/fregate-rs/blob/main/src/application/management.rs) configuration. Currently only endpoints configuration is supported.
@@ -27,6 +29,7 @@ pub struct ManagementConfig {
 /// const LIVE_ENDPOINT: &str = "/live";
 /// const READY_ENDPOINT: &str = "/ready";
 /// const METRICS_ENDPOINT: &str = "/metrics";
+/// const VERSION_ENDPOINT: &str = "/{component_name}/version";
 /// ```
 /// You might want to change those:\
 /// Example:
@@ -68,6 +71,8 @@ pub struct Endpoints {
     pub ready: Endpoint,
     /// metrics endpoint
     pub metrics: Endpoint,
+    /// version endpoint
+    pub version: Endpoint,
 }
 
 #[allow(clippy::indexing_slicing)]
@@ -80,6 +85,7 @@ impl<'de> Deserialize<'de> for Endpoints {
         static_assert!(LIVE_ENDPOINT.as_bytes()[0] == b'/');
         static_assert!(READY_ENDPOINT.as_bytes()[0] == b'/');
         static_assert!(METRICS_ENDPOINT.as_bytes()[0] == b'/');
+        static_assert!(VERSION_ENDPOINT.as_bytes()[0] == b'/');
 
         let value = Value::deserialize(deserializer)?;
 
@@ -95,12 +101,16 @@ impl<'de> Deserialize<'de> for Endpoints {
         let metrics = value
             .pointer_and_deserialize::<_, D::Error>(METRICS_PTR)
             .unwrap_or_else(|_| Endpoint(METRICS_ENDPOINT.to_owned()));
+        let version = value
+            .pointer_and_deserialize::<_, D::Error>(VERSION_PTR)
+            .unwrap_or_else(|_| Endpoint(VERSION_ENDPOINT.to_owned()));
 
         Ok(Endpoints {
             health,
             live,
             ready,
             metrics,
+            version,
         })
     }
 }
@@ -112,12 +122,14 @@ impl Default for Endpoints {
         static_assert!(LIVE_ENDPOINT.as_bytes()[0] == b'/');
         static_assert!(READY_ENDPOINT.as_bytes()[0] == b'/');
         static_assert!(METRICS_ENDPOINT.as_bytes()[0] == b'/');
+        static_assert!(VERSION_ENDPOINT.as_bytes()[0] == b'/');
 
         Self {
             health: Endpoint(HEALTH_ENDPOINT.to_owned()),
             live: Endpoint(LIVE_ENDPOINT.to_owned()),
             ready: Endpoint(READY_ENDPOINT.to_owned()),
             metrics: Endpoint(METRICS_ENDPOINT.to_owned()),
+            version: Endpoint(VERSION_ENDPOINT.to_owned()),
         }
     }
 }