+++ tests/bootstrap.rs
@@ -13,9 +13,9 @@ mod bootstrap_fn_test {
 
         let logger = &config.observability_cfg;
         assert_eq!(logger.traces_endpoint, None);
-        assert_eq!(logger.service_name, "fregate".to_owned());
-        assert_eq!(logger.version, "0.1.0".to_owned());
-        assert_eq!(logger.component_name, "example".to_owned());
+        assert_eq!(logger.service_name, "default".to_owned());
+        assert_eq!(logger.version, "default".to_owned());
+        assert_eq!(logger.component_name, "default".to_owned());
         assert_eq!(logger.trace_level, "info".to_owned());
         assert_eq!(logger.log_level, "info".to_owned());